+++ Develop/Server/GameServer/main/GQuestReseter.cpp
@@ -2,7 +2,7 @@
 #include "GQuestReseter.h"
 #include "GEntityPlayer.h"
 #include "GPlayerQuests.h"
-#include "CCommandTable.h"
+#include "CCommandTable_GM.h"
 #include "GCommand.h"
 #include "GGlobal.h"
 #include "GDBManager.h"
@@ -15,7 +15,7 @@ bool GQuestReseter::ResetDoneQuestForGM( GEntityPlayer* pPlayer )
 
 	pPlayer->GetQuests().ClearDoneCount();
 
-	gsys.pDBManager->GM_QUEST_HISTORY_RESET_ALL(pPlayer->GetAID(), (int64)pPlayer->GetCID());
+	gsys.pDBManager->GM_QUEST_HISTORY_RESET_ALL(pPlayer->GetAID(), pPlayer->GetCID());
 
 	MCommand* pNewCmd = MakeNewCommand(MC_GM_QUEST_RESET, 0, NULL);
 
@@ -32,9 +32,9 @@ bool GQuestReseter::ResetAllQuestForGM( GEntityPlayer* pPlayer )
 	pPlayer->GetQuests().Clear();
 	pPlayer->GetQuests().ClearDoneCount();	
 
-	gsys.pDBManager->GM_QUEST_RESET_ALL(pPlayer->GetAID(), (int64)pPlayer->GetCID());
+	gsys.pDBManager->GM_QUEST_RESET_ALL(pPlayer->GetAID(), pPlayer->GetCID());
 
-	gsys.pDBManager->GM_QUEST_HISTORY_RESET_ALL(pPlayer->GetAID(), (int64)pPlayer->GetCID());
+	gsys.pDBManager->GM_QUEST_HISTORY_RESET_ALL(pPlayer->GetAID(), pPlayer->GetCID());
 
 	MCommand* pNewCmd = MakeNewCommand(MC_GM_QUEST_RESET, 0, NULL);
